requirements.txt
channels==2.4.0
channels-redis==2.3.2
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
redis==3.4.1